https://github.com/StockSharp/StockSharp/pull/175 refactoring.
